smaz_implementations-base_64: rewrite variable-length verbatim code
At some point in the development I changed the base-64 scheme, but failed
to propagate it everywhere. Mostly I intend to use only one mode in
base-64, and the other one is merely a place-holder, which explains how
it got neglected. Now at least it works, even if it's useless.